# apps/backend/core-api/docs/swagger.yaml
definitions:
...
  entity.InboxMessageType:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - InboxMessageTypeGeneral
    - InboxMessageTypeEventRegistrationInvitation
    - InboxMessageTypeEventCertificateInvitation
...
  inbox.InboxMessageCertificateInvitationViewModel:
    properties:
      certificate_id:
        type: string
      certificate_title:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      event_id:
        type: string
      event_name:
        type: string
      has_participant_joined_event:
        type: boolean
      hidden_at:
        type: string
      id:
        type: string
      is_read:
        type: integer
      message_content:
        type: string
      message_content_fallback:
        type: string
      message_type:
        $ref: '#/definitions/entity.InboxMessageType'
      receiver_email:
        type: string
      receiver_wallet_address:
        type: string
      revoked_at:
        type: string
      sender_credential_email:
        type: string
      sender_credential_wallet_address:
        type: string
      token_id:
        type: string
      updated_at:
        type: string
    required:
    - certificate_id
    - created_at
    - created_at
    - event_id
    - event_name
    - has_participant_joined_event
    - id
    - is_read
    - message_content
    - message_type
    - updated_at
    type: object
  inbox.InboxMessagesEventRegistrationInvitationViewModel:
    properties:
      academic_institution:
        type: string
      accepted_at:
        type: string
      broadcasted_at:
        type: string
      cancelled_at:
        type: string
      certificate_id:
        type: string
      certificate_title:
        type: string
      code:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      event_id:
        type: string
      event_name:
        type: string
      first_name:
        type: string
      has_participant_joined_event:
        type: boolean
      hidden_at:
        type: string
      id:
        type: string
      is_read:
        type: integer
      last_name:
        type: string
      message_content:
        type: string
      message_content_fallback:
        type: string
      message_type:
        $ref: '#/definitions/entity.InboxMessageType'
      phone_number:
        type: string
      receiver_email:
        type: string
      receiver_wallet_address:
        type: string
      sender_credential_email:
        type: string
      sender_credential_wallet_address:
        type: string
      token_id:
        type: string
      updated_at:
        type: string
      valid_until:
        type: string
    required:
    - created_at
    - created_at
    - event_id
    - id
    - is_read
    - message_content
    - message_type
    - updated_at
    - updated_at
    type: object
  inbox.InboxMessagesViewModel:
    properties:
      certificate_id:
        type: string
      certificate_title:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      event_id:
        description: Certificate-specific fields (only populated for certificate invitation
          messages)
        type: string
      event_name:
        type: string
      has_participant_joined_event:
        type: boolean
      hidden_at:
        type: string
      id:
        type: string
      is_read:
        type: integer
      message_content:
        type: string
      message_content_fallback:
        type: string
      message_type:
        $ref: '#/definitions/entity.InboxMessageType'
      receiver_email:
        type: string
      receiver_wallet_address:
        type: string
      sender_credential_email:
        type: string
      sender_credential_wallet_address:
        type: string
      token_id:
        type: string
      updated_at:
        type: string
    required:
    - created_at
    - id
    - is_read
    - message_content
    - message_type
    - updated_at
    type: object
  inboxmessages.GetInboxMessageResponse:
    properties:
      event_certificate:
        $ref: '#/definitions/inbox.InboxMessageCertificateInvitationViewModel'
      event_registration_invitation:
        $ref: '#/definitions/inbox.InboxMessagesEventRegistrationInvitationViewModel'
      inbox_message:
        $ref: '#/definitions/inbox.InboxMessagesViewModel'
      inbox_message_type:
        $ref: '#/definitions/entity.InboxMessageType'
    required:
    - inbox_message
    - inbox_message_type
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: DECM (Decentralized Event Management) platform API for NFT ticketing,
    digital credentials, and academic identity verification.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: DECM Core
  version: "1.0"
paths:
...
  /api/v1/inbox-messages/{inbox_message_id}:
    get:
      consumes:
      - application/json
      description: Get inbox message
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/inboxmessages.GetInboxMessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Get inbox message
      tags:
      - Inbox Messages
...
swagger: "2.0"